final-projects/CPSC-44000/Albert-Final-Project-main/public/azure-function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.25.0
Pillow==8.0.1
